Revert "dts: somc: sm8150: Disallow fsa4480 from using GPIO100"
This reverts commit 064aa67018a4eaff88ea286be5e5f624949f8b3f.

It is a wiser choice to keep many of kumano's modules as loadable instead
of built-in due to boot partition size restrictions.

If we are going down that route, though, that means that the GPIO100 clash
would not interfere with the panel at boot time anymore: therefore, revert
this patch as it is now useless.

Signed-off-by: voidanix <voidanix@keyedlimepie.org>